% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{peaks}
\alias{peaks}
\title{peaks}
\usage{
peaks(x, span = NULL, partial = TRUE, height = NULL, smooth.method =
  NULL, smooth.n = NULL, plot = TRUE, plot.which = "b")
}
\arguments{
\item{x}{numeric. A numeric vector.}

\item{span}{integer. Sets window size for peak (or trough) identification; to
be designated a peak, a value must have \code{span} \emph{lower} values on
\emph{both} sides of it (higher for troughs).}

\item{partial}{logical. Default TRUE. Should the function attempt to identify
peaks or troughs at the start or end of the vector where \code{span} is
truncated.}

\item{height}{numeric. Value between 0 and 1. Sets threshold for peak or
trough 'prominence'. See Details.}

\item{smooth.method}{string. Method by which to smooth data before peak
identification. Optional, default is \code{NULL}. See \code{\link[=smooth]{smooth()}}.}

\item{smooth.n}{string. Smoothing factor. See \code{\link[=smooth]{smooth()}}.}

\item{plot}{logical. Plots the result.}

\item{plot.which}{string. What to plot: "p" for peaks, "t" for troughs, or
the default "b" for both.}
}
\description{
Identify peaks and troughs in oscillating data.
}
\details{
\code{peaks} scans a vector of numeric values and identifies peaks and
troughs.

The required factor (\code{span}) and two optional factors (smoothing and
\code{height}), need to be balanced to successfully identify peaks and troughs.
Characteristics such as data noisiness, amplitude, etc will affect how
important these are and how successful the identification process is.

\strong{Span}

The most important parameter in determining a peak is the \code{span}. This
sets the threshold for identification; to be designated a peak, a value
(after smoothing) must be the highest value within the \code{span} window
(lowest value for troughs).

\strong{Smoothing}

For noisy data there is optional smoothing functionality via
\code{smooth_method}. See \code{\link[=smooth]{smooth()}} for the methods available and appropriate
\code{smooth.n} values. \code{smooth.method = "spline"} works particularly well for
oscillating data.

\strong{Height}

Peak identification is also affected by the optional \code{height} argument.
This sets a threshold of 'prominence' for peak identification (or 'depth'
for a trough). This should be a value between 0 and 1 describing a
proportion of the total \code{range} of the data. Only peaks or troughs
containing a proportional range equal to or above this within their \code{span}
window will be retained. Essentially the higher this is set, the more
prominent the peaks, or deeper the troughs, must be to be identified. To
help with appropriate values, range elements are included in the output.
Peaks and troughs can have different \code{height} thresholds: simply enter them
as a vector of two values in peak-trough order. E.g. \code{height = c(0.2, 0.1)}.

\strong{Matching values}

If there happens to be equal values within the \code{span} range, the first
occurrence is designated as the peak. If peaks are being missed because of
this, use a lower \code{span}. If there are many instances of equal values, try
smoothing the data.

\strong{Partial windows}

A rolling window of width \code{span} is used across the data to identify peaks.
At the start and end of the data vector this window will overlap the start
and end. The default \code{partial = TRUE} input tells the function to attempt
to identify peaks near the start/end of the data where the \code{span} input
would not supply enough data points. Change this to \code{FALSE} if you see odd
matching behaviour at the start or end of the data.

\strong{Plot}

If \code{plot = TRUE} a plot is produced of the data with peaks and/or troughs
highlighted. If the data has been smoothed this will be a 2-panel plot of
both data sets, otherwise it is the raw data. \code{plot.which} controls what
results are plotted. The default \code{plot.which = "b"} means \emph{both} peaks
(green points and lines) and troughs (blue points and lines) are
highlighted within the data. Change to \code{plot.which = "p"} to plot only
peaks, or \code{plot.which = "t"} for only troughs. This does not affect the
output.

\strong{Output}

The output of the function is a \code{list} object of class \code{peaks} containing
the following elements which can be extracted via the \code{$} operator for
further use:
\itemize{
\item \code{data} - original data input
\item \code{data_smooth} - smoothed data. Identical to \code{data} if no smoothing
applied
\item \code{peaks} - index of locations of identified peaks
\item \code{troughs} - index of locations of identified troughs
\item \code{both} - index of locations of \emph{both} peaks and troughs
\item \code{peaks_range} - proportional data range within the \code{span} window of each
peak
\item \code{troughs_range} - proportional data range within the \code{span} window of
each trough
\item \code{both_range} - proportional data range within the \code{span} window of each
peak and trough
\item \code{call} - the function call
}
}
\examples{

}
\author{
Nicholas Carey - \email{nicholascarey@gmail.com}
}
