% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{Display image generation progress bar}
\usage{
progress(x, max = 100)
}
\description{
Display image generation progress bar
}
\details{
Displays an updating progress bar in the console. Modified from
\url{https://github.com/klmr/rcane/blob/master/system.R}.
}
