% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_head.R
\name{replace_head}
\alias{replace_head}
\title{replace_head}
\usage{
replace_head(v, n, r)
}
\arguments{
\item{v}{vector. Input vector to have head replaced.}

\item{n}{integer. Number of values at start of \code{v} to replace. If left NULL,
this will be calculated as the length of \code{r}.}

\item{r}{String, logical, numeric etc. What to use as replacement. Should
either be a single value or vector of length \code{n}.}
}
\value{
The \code{v} vector is returned but with the initial \code{n} values overwritten
by `r.
}
\description{
Replace initial \code{n} values in a vector.
}
\details{
A simple solution for replacing the initial \code{n} values in a vector.
Replacing the initial \code{n} values is straightforward compared to replacing
the final \code{n} values, but this makes a nice partner function to
\code{\link[=replace_head]{replace_head()}}.
}
\examples{
# Replace initial 5 numeric values with single value
x <- 1:10
replace_head(x, 5, 100)

# Replace initial logical with a different logical
x <- rep(TRUE, 10)
replace_head(x, 1, FALSE)

# Replace initial n with a vector of same length
x <- 1:20
replace_head(x, 5, c(100,99,98,97,96))

}
\seealso{
replace_head
}
\author{
Nicholas Carey - \email{nicholascarey@gmail.com}
}
