% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_tail.R
\name{replace_tail}
\alias{replace_tail}
\title{replace_tail}
\usage{
replace_tail(v, n, r)
}
\arguments{
\item{v}{vector. Input vector to have tail replaced.}

\item{n}{integer. Number of values at end of \code{v} to replace.}

\item{r}{String, logical, numeric etc. What to use as replacement. Should
either be a single value or vector of length \code{n}.}
}
\value{
The \code{v} vector is returned but with the final \code{n} values overwritten
by \code{r}.
}
\description{
Replace final \code{n} values in a vector.
}
\details{
A simple solution to the lack of an elegant way in \code{R} of simply
replacing the last \code{n} values in a vector. There are lots of ways of
extracting the last \code{n} values, and a few of these can be used to replace
them, but they are all not very elegant in code. There is a similar
function for the much easier operation to replace the initial \code{n} values:
\code{\link[=replace_head]{replace_head()}}.
}
\examples{
# Replace last 5 numeric values with single value
x <- 1:10
replace_tail(x, 5, 100)

# Replace last logical with a different logical
x <- rep(TRUE, 10)
replace_tail(x, 1, FALSE)

# Replace last n with a vector of same length
x <- 1:20
replace_tail(x, 5, c(100,99,98,97,96))

}
\seealso{
replace_head
}
\author{
Nicholas Carey - \email{nicholascarey@gmail.com}
}
