% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resp_noisy.rd}
\alias{resp_noisy.rd}
\title{Noisy respirometry data}
\format{
A vector of 935 values
}
\usage{
resp_noisy.rd
}
\description{
Comes from a flowthrough respirometry experiment. For testing smoothing
function.
}
\author{
Nicholas Carey
}
\keyword{datasets}
