% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right.R
\name{right}
\alias{right}
\title{Right substring function.}
\usage{
right(x, n)
}
\arguments{
\item{x}{numeric value or string.}

\item{n}{numeric. The number of characters to extract from the right side.}
}
\description{
Takes a string or numeric value and subsets out the right n characters.
\subsection{Details}{

A simple function that extracts the \emph{right} number of specified characters
from a string or numeric value.
}

\subsection{Output}{

The returned output will be same class as the input.
}
}
\examples{
right("some text", 4)
right(123456, 4)

}
\seealso{
\code{\link{left}}
}
\author{
Nicholas Carey - \email{nicholascarey@gmail.com}
}
