% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sine_noisy.rd}
\alias{sine_noisy.rd}
\title{Slightly noisy sine wave type data}
\format{
A vector of 96 values
}
\usage{
sine_noisy.rd
}
\description{
For testing smoothing and peak finding functions.
}
\author{
Nicholas Carey
}
\keyword{datasets}
