% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swim_y.rd}
\alias{swim_y.rd}
\title{Spatial y-coordinates of a fish tail during swimming}
\format{
A vector of 266 values
}
\usage{
swim_y.rd
}
\description{
For testing smoothing and peak finding functions.
}
\author{
Nicholas Carey
}
\keyword{datasets}
